Fix final build

+++ database/migrations/2020_10_21_171837_create_orders_table.php
@@ -28,7 +28,7 @@ class CreateOrdersTable extends Migration
             $table->decimal('promocode_discount')->default(0);
             $table->text('comment')->default('');
             $table->boolean('viewed')->default(0);
-            $table->unsignedInteger('manager_id')->nullable();
+            $table->unsignedBigInteger('manager_id')->nullable();
             $table->timestamps();
 
             $table->foreign('status_id')->references('order_status_id')->on('order_statuses');
