+++ database/migrations/2022_03_31_193906_add_order_history_table.php
@@ -19,7 +19,7 @@ class AddOrderHistoryTable extends Migration
             $table->boolean('notify_customer')->default(0);
             $table->string('history_comment')->nullable();
             $table->string('history_status');
-            $table->unsignedInteger('manager_id')->nullable();
+            $table->unsignedBigInteger('manager_id')->nullable();
             $table->timestamps();
 
             $table->foreign('order_id')->references('order_id')->on('orders');
